refactor: split oversized JSX files into focused components
App.jsx (1022→436): Extract 8 modules:
- hooks/useServerHealth.js, useNavStack.js, useRunNavigator.js
- constants/navigation.jsx (icon SVGs)
- components/ServerDisconnectedOverlay.jsx, Sidebar.jsx, ProjectHeader.jsx
- features/evaluation/components/EvaluateScreen.jsx
- features/settings/components/SettingsPage.jsx

DashboardPage.jsx (701→122): Extract 2 panel components:
- AccumulatedOverviewPanel.jsx (329 lines)
- RunOverviewPanel.jsx (288 lines)
