+++ ui/web/src/constants/navigation.jsx
@@ -1,0 +1,38 @@
+// ---------------------------------------------------------------------------
+// Sidebar icon constants and navigation items
+// ---------------------------------------------------------------------------
+
+export const ICON_OVERVIEW = (
+  <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+    <rect x="3" y="13" width="4" height="8" />
+    <rect x="10" y="8" width="4" height="13" />
+    <rect x="17" y="3" width="4" height="18" />
+  </svg>
+);
+
+export const ICON_EVALUATE = (
+  <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+    <circle cx="11" cy="11" r="7" />
+    <line x1="16.5" y1="16.5" x2="21" y2="21" />
+    <line x1="8" y1="11" x2="14" y2="11" />
+    <line x1="11" y1="8" x2="11" y2="14" />
+  </svg>
+);
+
+export const ICON_PROJECTS = (
+  <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+    <polyline points="3,4 5.5,6 3,8" />
+    <line x1="9" y1="6" x2="21" y2="6" />
+    <polyline points="3,11 5.5,13 3,15" />
+    <line x1="9" y1="13" x2="21" y2="13" />
+    <polyline points="3,18 5.5,20 3,22" />
+    <line x1="9" y1="20" x2="21" y2="20" />
+  </svg>
+);
+
+export const ICON_SETTINGS = (
+  <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
+    <circle cx="12" cy="12" r="3" />
+    <path d="M12 2v2.5M12 19.5V22M4.93 4.93l1.77 1.77M17.3 17.3l1.77 1.77M2 12h2.5M19.5 12H22M4.93 19.07l1.77-1.77M17.3 6.7l1.77-1.77" />
+  </svg>
+);
